Composing of the Item's date was refactored and moved into shared service.
